:hammer: Move Header component to global components folder
